// src/app/models/admin/dangkyhoatdongkhoangsan/dkkhaithackhoangsanduan.model.ts
export class OutputDkKhaiThacKhoangSanDuAnModel {
    iddangkykhaithac: string;
    tenduan: string;
    soquyetdinh: string;
    ngaykyquyetdinh: string;
    coquanpheduyet: string;
    diadiem: string;
    dientichkhaithac: number;
    truluongdiachat: number;
    truluongkhaithac: number;
    thoihankhaithac: number;
    phuongphapkhaithac: number;
    congsuatkhaithac: number;
    mucsaukhaithactu: number;
    mucsaukhaithacden: number;
    dangkhoangsan: number;
    mucdichsudungkhoangsan: string;
    donvitruluong: string;
    donvicongsuat: string;
    donvidientich: string;
    donvithoihan: string;
    donvichieusau: string;
    geom: string;
    geovn: string;
    geowgs: string;
    hequychieu: string;
    idhoso: string;
    siteid: string
}

export class InputDkKhaiThacKhoangSanDuAnModel {
    iddangkykhaithac: string;
    tenduan: string;
    soquyetdinh: string;
    ngaykyquyetdinh: string;
    coquanpheduyet: string;
    diadiem: string;
    dientichkhaithac: number;
    truluongdiachat: number;
    truluongkhaithac: number;
    thoihankhaithac: number;
    phuongphapkhaithac: number;
    congsuatkhaithac: number;
    mucsaukhaithactu: number;
    mucsaukhaithacden: number;
    dangkhoangsan: number;
    mucdichsudungkhoangsan: string;
    donvitruluong: string;
    donvicongsuat: string;
    donvidientich: string;
    donvithoihan: string;
    donvichieusau: string;
    hequychieu: string;
    idhoso: string;
}